requirements.txt
qdrant-client
sentence-transformers
python-dotenv
python-frontmatter
flask
